src/main/java/com/example/chatmessagner/service/UserService.java
package com.example.chatmessagner.service;

import com.example.chatmessagner.entity.User;
import com.example.chatmessagner.repository.UserRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.security.provisioning.UserDetailsManager;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.Optional;


@Service
@Slf4j
public class UserService implements UserDetailsManager {


    private final UserRepository userRepository;

    private final BCryptPasswordEncoder bCryptPasswordEncoder;


    @Autowired
    public UserService(UserRepository userRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
        this.userRepository = userRepository;
        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
    }

    @Override
    public void createUser(UserDetails user) {

        User user0 = (User) user;

        user0.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));

        userRepository.save((User) user);
    }

    @Override
    public void updateUser(UserDetails user) {
        userRepository.save((User) user);
    }

    @Override
    public void deleteUser(String phoneNumber) {
        userRepository.deleteByPhoneNumber(phoneNumber);
    }

    @Override
    public void changePassword(String oldPassword, String newPassword) {
        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();

        user.setPassword(bCryptPasswordEncoder.encode(newPassword));
        userRepository.save(user);
    }

    @Override
    public boolean userExists(String phoneNumber) {
        return userRepository.findByPhoneNumber(phoneNumber).isPresent();
    }

    @Override
    public UserDetails loadUserByUsername(String phoneNumber) throws UsernameNotFoundException {
        return userRepository.findByPhoneNumber(phoneNumber)
                .orElseThrow(() -> new UsernameNotFoundException("User not found by phoneNumber: " + phoneNumber));
    }

    public Optional<User> getById(Integer id) {
        return userRepository.findById(id);
    }

    @Transactional
    public void addContact(User user, String phoneNumber) {
        userRepository.findByPhoneNumber(phoneNumber).ifPresent(user2 -> {
            userRepository.addContact(user.getId(), user2.getId());
        });
    }

    public List<User> getContacts(Integer userId) {
        log.info("Getting contacts not from cache...");
        return userRepository.findAllContacts(userId);
    }
}
